Fixed the default int32 type which was wrong on amd64 (and added testcase). Also, added an Ogg encoder (doesn't work quite yet).

+++ libcelt/celt_types.h
@@ -158,8 +158,8 @@
    /* Give up, take a reasonable guess */
    typedef short celt_int16_t;
    typedef unsigned short celt_uint16_t;
-   typedef long celt_int32_t;
-   typedef unsigned long celt_uint32_t;
+   typedef int celt_int32_t;
+   typedef unsigned int celt_uint32_t;
    typedef long long celt_int64_t;
    typedef unsigned long long celt_uint64_t;
